continued with styling for desktop and tablet, with issues i will bring up in townhall
