--- src/UILib.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-export const MainParagraph = styled.p`
-  font-size: 16px;
-  font-weight: 400;
-
-  @media (min-width: 1200px) {
-    font-size: 18px;
-  }
-`;
-
-/*library for storiing styled components*/
