--- src/sections/Desc.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const Desc = ({ desc }) => {
-  return <p className="mb-[32px] text-16 font-normal">{desc}</p>;
-};
