+++ src/sections/DescComponent/Desc.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const Desc = ({ desc }) => {
+  return <p className="mb-[32px] text-16 font-normal xl:ml[125px]">{desc}</p>;
+};
