+++ src/sections/FeaturedProjectsComponents/ProjectBox.jsx
@@ -4,7 +4,7 @@ import projects from "../../data/projects.json";
 
 import { Image } from "./ProjectBoxComponents/Image";
 import { Tags } from "./ProjectBoxComponents/Tags";
-import { Desc } from "../Desc";
+import { Desc } from "../DescComponent/Desc";
 import { Links } from "./ProjectBoxComponents/Links";
 import { Title } from "../TitleComponents/TitleH3";
 
@@ -12,14 +12,19 @@ export const ProjectBox = () => {
   return (
     <>
       {projects.projects.slice(0, 4).map((project) => (
-        <section key={project.id} className="mb-[64px]">
+        <section
+          key={project.id}
+          className="mb-[64px] xl:mb-[128px] xl:flex xl:items-center"
+        >
           <Image url={project.image} />
-          <Tags tags={project.tags} />
-          <article>
-            <Title title={project.name} />
-            <Desc desc={project.desc} />
-          </article>
-          <Links netlf={project.netlify} github={project.github} />
+          <section className="xl:ml-[125px] xl:mr-[128px]">
+            <Tags tags={project.tags} />
+            <article>
+              <Title title={project.name} />
+              <Desc desc={project.desc} />
+            </article>
+            <Links netlf={project.netlify} github={project.github} />
+          </section>
         </section>
       ))}
     </>
