+++ src/sections/TitleComponents/Title.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 export const Title = ({ title }) => {
-  return <h2 className="text-center pb-[64px]">{title}</h2>;
+  return <h2 className="text-center pb-[64px] xl:pb-[128px] ">{title}</h2>;
 };
 
 //mb-3??
